backend/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
drf-yasg==1.20.0
django-cors-headers==3.6.0
celery==5.0.5
redis==3.5.3
